Refactor the whole feature to be Linux-specific
Trying to have a separation between generic
CgroupResourceUsageSamplingRunner and Linux-specific Cgroup reader
started to feel off, let's make the whole feature Linux-only.
