+++ pkg/runner/BUILD.bazel
@@ -4,9 +4,8 @@ go_library(
     name = "runner",
     srcs = [
         "apple_xcode_resolving_runner.go",
-        "cgroup_linux.go",
-        "cgroup_other.go",
-        "cgroup_resource_usage_sampling_runner.go",
+        "cgroup_resource_usage_sampling_runner_linux.go",
+        "cgroup_resource_usage_sampling_runner_other.go",
         "clean_runner.go",
         "local_runner.go",
         "local_runner_darwin.go",
@@ -22,7 +21,6 @@ go_library(
     visibility = ["//visibility:public"],
     deps = [
         "//pkg/cleaner",
-        "//pkg/proto/resourceusage",
         "//pkg/proto/runner",
         "//pkg/proto/tmp_installer",
         "@com_github_buildbarn_bb_storage//pkg/filesystem",
@@ -35,26 +33,32 @@ go_library(
         "@org_golang_google_protobuf//types/known/emptypb",
     ] + select({
         "@rules_go//go/platform:android": [
+            "//pkg/proto/resourceusage",
             "@org_golang_google_protobuf//types/known/durationpb",
             "@org_golang_x_sys//unix",
         ],
         "@rules_go//go/platform:darwin": [
+            "//pkg/proto/resourceusage",
             "@org_golang_google_protobuf//types/known/durationpb",
             "@org_golang_x_sys//unix",
         ],
         "@rules_go//go/platform:freebsd": [
+            "//pkg/proto/resourceusage",
             "@org_golang_google_protobuf//types/known/durationpb",
             "@org_golang_x_sys//unix",
         ],
         "@rules_go//go/platform:ios": [
+            "//pkg/proto/resourceusage",
             "@org_golang_google_protobuf//types/known/durationpb",
             "@org_golang_x_sys//unix",
         ],
         "@rules_go//go/platform:linux": [
+            "//pkg/proto/resourceusage",
             "@org_golang_google_protobuf//types/known/durationpb",
             "@org_golang_x_sys//unix",
         ],
         "@rules_go//go/platform:windows": [
+            "//pkg/proto/resourceusage",
             "@org_golang_google_protobuf//types/known/durationpb",
             "@org_golang_x_sys//windows",
         ],
@@ -72,11 +76,9 @@ go_test(
         "temporary_directory_symlinking_runner_test.go",
     ] + select({
         "@rules_go//go/platform:android": [
-            "cgroup_linux_test.go",
             "cgroup_resource_usage_sampling_runner_test.go",
         ],
         "@rules_go//go/platform:linux": [
-            "cgroup_linux_test.go",
             "cgroup_resource_usage_sampling_runner_test.go",
         ],
         "//conditions:default": [],
